improved mesher to account for absorbing BC
Former-commit-id: cab833803fb7acabfca8346fc9ccae7868be865b

+++ src/force_external.py
@@ -13,7 +13,7 @@ class Force:
         from scipy.sparse import lil_matrix
 
         time = load_set["time"]
-        time = np.linspace(0, time, np.ceil(time / time_step))
+        time = np.linspace(0, time, int(np.ceil(time / time_step)))
         # generation of variable
         self.force = lil_matrix((nb_equations, len(time)))
